move plugin management service and models to '@valtimo/plugin'
