+++ projects/valtimo/form-link/src/lib/components/select-plugin-function/select-plugin-function.component.ts
@@ -21,7 +21,7 @@ import {
   PluginConfiguration,
   PluginFunction,
   PluginManagementService,
-} from '@valtimo/plugin-management';
+} from '@valtimo/plugin';
 import {ProcessLinkStateService} from '../../services/process-link-state.service';
 import {Observable, of} from 'rxjs';
