+++ projects/valtimo/plugin-management/src/lib/services/index.ts
@@ -14,5 +14,4 @@
  * limitations under the License.
  */
 
-export * from './plugin-management.service';
 export * from './plugin-management-state.service';
