+++ projects/valtimo/plugin/src/lib/services/index.ts
@@ -16,3 +16,4 @@
 
 export * from './plugin.service';
 export * from './plugin-translation.service';
+export * from './plugin-management.service';
